app/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
pytest-django==3.9.0
pytest==5.4.1
pytest-cov==2.8.1
flake8==3.7.9
coverage==5.1
black==19.10b0
isort==4.3.21
Pillow==6.2.1
django-storages==1.9.1
boto3==1.12.49
celery==4.4.1
redis==3.4.1